refactor(governance)!: extract seam contracts into seams package
Extract NormativeProvider, AttestationProvider, and ExecutionActuator seam
contracts into dedicated src/gateway/governance/seams/ package to eliminate
circular dependencies with vendor adapters.

Changes:
- Created seams/normative.py with NormativeBaseline, ValidationResult,
  EvidenceSeal, ExecutionStatus, FindingStatus, NormativeProvider protocol
- Created seams/attestation.py with AttestationProvider, AttestationStatus,
  ExternalAttestation
- Created seams/actuation.py with ExecutionActuator, ExecutionClearance,
  ActuationReceipt, ActuatorCapability
- Updated normative_provider.py, attestation_provider.py, execution_actuator.py,
  governance_envelope.py to import from seams/
- Converted provider_01, provider_03, provider_06 from function-scope to
  module-scope seam imports (removed 12 function-scope imports)
- Updated all test imports to use seam modules
- Left provider_02 untouched per C1 specification

Architectural invariant: Seam modules have ZERO imports from the kernel.
This is the load-bearing property that breaks the circular dependency.

Gate G3 allowlist (check_import_boundaries.py) unchanged — function-scope
imports remain in normative_provider.py and evidence/factory.py for vendor
adapter lazy-loading.

BREAKING CHANGE: Seam contract imports must now use
src.gateway.governance.seams.{normative,attestation,actuation} instead of
importing from normative_provider.py, attestation_provider.py,
execution_actuator.py, or governance_envelope.py. The old modules now
re-export seam contracts but the re-exports are removed in this same commit
per C0 specification (no deprecation window for reference architecture).

Test results: 3839 passed, 99 skipped (baseline preserved)
Import boundaries: ✅ All boundaries respected (Gate G3)
Mypy: 1 pre-existing error in provider_02 (not ours to fix, folded into C1)
