+++ scripts/test_gke_e2e_flow.py
@@ -3,35 +3,38 @@
 End-to-End Functional Test for Live GKE Deployment
 Tests actual governance flow through the gateway.
 """
+
+import json
 import os
 import sys
-import json
-import httpx
 import time
+
+import httpx
+
 
 def test_gateway_governance_flow():
     """Test a simple governance flow through the gateway"""
     print("\n=== Testing Gateway Governance Flow ===")
-    
+
     gateway_url = "http://localhost:8080"
-    
+
     # Test 1: Health check
     print("\n1. Gateway health check...")
     try:
         response = httpx.get(f"{gateway_url}/health", timeout=10)
-        assert response.status_code == 200, f"Health check failed: {response.status_code}"
+        assert response.status_code == 200, (
+            f"Health check failed: {response.status_code}"
+        )
         print(f"✓ Health check passed: {response.json()}")
     except Exception as e:
         print(f"✗ Health check failed: {e}")
         return False
-    
+
     # Test 2: MCP tools listing
     print("\n2. MCP tools listing...")
     try:
         response = httpx.post(
-            f"{gateway_url}/mcp",
-            json={"method": "tools/list"},
-            timeout=10
+            f"{gateway_url}/mcp", json={"method": "tools/list"}, timeout=10
         )
         if response.status_code == 200:
             data = response.json()
@@ -43,14 +46,11 @@ def test_gateway_governance_flow():
             print(f"  Note: MCP endpoint returned {response.status_code}")
     except Exception as e:
         print(f"  Note: MCP test skipped: {e}")
-    
+
     # Test 3: OPA policy evaluation (via compliance bridge)
     print("\n3. Compliance bridge control check...")
     try:
-        response = httpx.get(
-            "http://localhost:3002/v1/controls",
-            timeout=10
-        )
+        response = httpx.get("http://localhost:3002/v1/controls", timeout=10)
         if response.status_code == 200:
             controls = response.json()
             print(f"✓ Compliance controls available: {len(controls)} controls")
@@ -59,13 +59,14 @@ def test_gateway_governance_flow():
             print(f"  Status: {response.status_code}")
     except Exception as e:
         print(f"  Note: Compliance check: {e}")
-    
+
     return True
+
 
 def test_vllm_inference():
     """Test vLLM inference endpoint"""
     print("\n=== Testing vLLM Inference ===")
-    
+
     # Test fast model
     print("\n1. vLLM Fast Model (Qwen 7B)...")
     try:
@@ -75,14 +76,14 @@ def test_vllm_inference():
                 "model": "gs://laah-cybernetics-models/models--Qwen--Qwen2.5-7B-Instruct/snapshots/a09a35458c702b33eeacc393d103063234e8bc28",
                 "prompt": "Hello, this is a test. Reply with 'OK' if you can read this.",
                 "max_tokens": 10,
-                "temperature": 0.1
+                "temperature": 0.1,
             },
-            timeout=30
+            timeout=30,
         )
         if response.status_code == 200:
             result = response.json()
             text = result.get("choices", [{}])[0].get("text", "").strip()
-            print(f"✓ vLLM Fast inference successful")
+            print("✓ vLLM Fast inference successful")
             print(f"  Response: {text[:100]}")
             return True
         else:
@@ -93,30 +94,25 @@ def test_vllm_inference():
         print(f"✗ vLLM inference failed: {e}")
         return False
 
+
 def test_langfuse_trace_creation():
     """Test Langfuse trace creation"""
     print("\n=== Testing Langfuse Trace Creation ===")
-    
+
     try:
         # Create a simple trace
         trace_id = f"test-trace-{int(time.time())}"
         response = httpx.post(
             "http://localhost:3001/api/public/traces",
-            auth=(
-                os.getenv("LANGFUSE_PUBLIC_KEY"),
-                os.getenv("LANGFUSE_SECRET_KEY")
-            ),
+            auth=(os.getenv("LANGFUSE_PUBLIC_KEY"), os.getenv("LANGFUSE_SECRET_KEY")),
             json={
                 "id": trace_id,
                 "name": "smoke-test",
-                "metadata": {
-                    "test": "gke-smoke-test",
-                    "timestamp": int(time.time())
-                }
+                "metadata": {"test": "gke-smoke-test", "timestamp": int(time.time())},
             },
-            timeout=10
+            timeout=10,
         )
-        
+
         if response.status_code in [200, 201]:
             print(f"✓ Trace created successfully: {trace_id}")
             return True
@@ -128,41 +124,43 @@ def test_langfuse_trace_creation():
         print(f"✗ Trace creation failed: {e}")
         return False
 
+
 def main():
     print("=" * 60)
     print("CAGE Live GKE End-to-End Functional Test")
     print("=" * 60)
-    
+
     results = {}
-    
+
     # Test gateway governance flow
     results["gateway_flow"] = test_gateway_governance_flow()
-    
+
     # Test vLLM inference
     results["vllm_inference"] = test_vllm_inference()
-    
+
     # Test Langfuse trace creation
     results["langfuse_trace"] = test_langfuse_trace_creation()
-    
+
     # Summary
     print("\n" + "=" * 60)
     print("FUNCTIONAL TEST SUMMARY")
     print("=" * 60)
     passed = sum(1 for v in results.values() if v)
     total = len(results)
-    
+
     for test, success in results.items():
         status = "✓ PASS" if success else "✗ FAIL"
         print(f"{status:8} {test}")
-    
+
     print(f"\nTotal: {passed}/{total} functional tests passed")
-    
+
     if passed == total:
         print("\n✓ All functional tests passed!")
         return 0
     else:
         print(f"\n⚠ {total - passed} test(s) failed (some may be expected)")
         return 0  # Return 0 for now as some tests may fail due to configuration
+
 
 if __name__ == "__main__":
     sys.exit(main())
